feature/task_6: executed a mapper in data layer , changed names ...ModelUseCase to UseCase , changed a body of function signIn in ImpEmployeesRepository.kt. Changed a minSDKVersion from 16 to 21
