+++ app/build.gradle
@@ -11,7 +11,7 @@ android {
     }
     defaultConfig {
         applicationId "com.example.loginapplication"
-        minSdkVersion 16
+        minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
