+++ app/src/main/java/Modules.kt
@@ -1,6 +1,5 @@
 import data.boundairds.IAuthService
 import data.boundairds.IEmployeesService
-import data.entity.Employee
 import data.entity.SignInResponse
 import data.repository.AuthRepository
 import data.repository.ImpEmployeesRepository
@@ -9,8 +8,8 @@ import data.services.StubEmployeesService
 import domain.boundaries.IAuthRepository
 import domain.boundaries.IEmployeesRepository
 import domain.entity.SignInResult
-import domain.usecases.EmployeesModelUseCase
-import domain.usecases.LoginModelUseCase
+import domain.usecases.EmployeesUseCase
+import domain.usecases.LoginUseCase
 import org.koin.dsl.module
 import presentation.LoginViewModel
 import presentation.Router
@@ -21,7 +20,7 @@ val appModule = module {
     single{SignInResponse(get())}
     single<IAuthService>{StubAuthServices()}
     single<IAuthRepository>{AuthRepository(get())}
-    single{LoginModelUseCase(get())}
+    single{LoginUseCase(get())}
     single{ Router() }
     single{LoginViewModel(get(),get(),get())}
     single{ToastManager(get())}
@@ -30,5 +29,5 @@ val appModule = module {
 val employeesModel = module {
     single<IEmployeesService>{StubEmployeesService()}
     single<IEmployeesRepository>{ImpEmployeesRepository(get())}
-    single{EmployeesModelUseCase(get())}
+    single{EmployeesUseCase(get())}
 }
