+++ app/src/main/java/data/entity/EmployeeMapper.kt
@@ -1,0 +1,10 @@
+package data.entity
+
+import domain.entity.EmployeeDTO
+
+fun fromEmployee(src: Employee) = EmployeeDTO(
+    src.getFirstName(),
+    src.getLastName(),
+    src.getUserName(),
+    src.getEmployeeData()
+)
