+++ app/src/main/java/data/repository/ImpEmployeesRepository.kt
@@ -2,16 +2,13 @@ package data.repository
 
 import data.boundairds.IEmployeesService
 import data.entity.Employee
+import data.entity.fromEmployee
 import domain.boundaries.IEmployeesRepository
 import domain.entity.EmployeeDTO
 
 class ImpEmployeesRepository(private val employeesService:IEmployeesService): IEmployeesRepository {
     override suspend fun getEmployes(): List<EmployeeDTO> {
         val listOfEmployee = employeesService.getEmployes()
-        val listOfEmployeeDTO:MutableList<EmployeeDTO> = mutableListOf()
-        listOfEmployee.forEach {
-            listOfEmployeeDTO.add(EmployeeDTO.fromEmployee(it))
-        }
-        return listOfEmployeeDTO
+        return listOfEmployee.map { employee-> fromEmployee(employee)}
     }
 }
