+++ app/src/main/java/domain/usecases/EmployeesUseCase.kt
@@ -3,7 +3,7 @@ package domain.usecases
 import domain.boundaries.IEmployeesRepository
 import domain.entity.EmployeeDTO
 
-class EmployeesModelUseCase(private val emplyessRepository: IEmployeesRepository) {
+class EmployeesUseCase(private val emplyessRepository: IEmployeesRepository) {
     suspend fun getEmployees(): List<EmployeeDTO> {
         return emplyessRepository.getEmployes()
     }
