+++ app/src/main/java/domain/usecases/LoginUseCase.kt
@@ -3,7 +3,7 @@ package domain.usecases
 import domain.boundaries.IAuthRepository
 import domain.entity.SignInResult
 
-class LoginModelUseCase(private val authRepository: IAuthRepository) {
+class LoginUseCase(private val authRepository: IAuthRepository) {
     suspend fun signIn(login:String, password:String): SignInResult
     {
         val response = authRepository.signIn(login,password)
